processor/requirements.txt
kafka-python-ng>=2.0.2
msgpack>=1.0.7
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
